Tests/LibWeb/Text/expected/HTML/StructuredClone-primitives.txt

undefined
null
true
false
123
123.456
9007199254740991

This is a string
